Qualify all implementations of data structures using the word "implementation."
