--- src/DataStructures/CustomDataStructures/LinkedList/DoublyCircularLinkedList/Node.java
@@ -1,14 +1,0 @@
-package DataStructures.CustomDataStructures.LinkedList.DoublyCircularLinkedList;
-
-public class Node<T> {
-
-    public T value;
-    public Node<T> previous;
-    public Node<T> next;
-
-    public Node(T newValue) {
-        this.value = newValue;
-        this.previous = null;
-        this.next = null;
-    }
-}
