+++ src/DataStructures/CustomDataStructures/LinkedList/DoublyCircularLinkedListImplementation/Node.java
@@ -1,0 +1,14 @@
+package DataStructures.CustomDataStructures.LinkedList.DoublyCircularLinkedListImplementation;
+
+public class Node<T> {
+
+    public T value;
+    public Node<T> previous;
+    public Node<T> next;
+
+    public Node(T newValue) {
+        this.value = newValue;
+        this.previous = null;
+        this.next = null;
+    }
+}
