+++ src/DataStructures/CustomDataStructures/LinkedList/DoublyLinkedListImplementation/Node.java
@@ -1,0 +1,14 @@
+package DataStructures.CustomDataStructures.LinkedList.DoublyLinkedListImplementation;
+
+class Node<T> {
+
+    public T value;
+    public Node<T> previous;
+    public Node<T> next;
+
+    public Node(T newValue) {
+        this.value = newValue;
+        this.previous = null;
+        this.next = null;
+    }
+}
