--- src/DataStructures/CustomDataStructures/LinkedList/SinglyCircularLinkedList/Node.java
@@ -1,12 +1,0 @@
-package DataStructures.CustomDataStructures.LinkedList.SinglyCircularLinkedList;
-
-public class Node<T> {
-
-    public T value;
-    public Node<T> next;
-
-    public Node(T newValue) {
-        this.value = newValue;
-        this.next = null;
-    }
-}
