+++ src/DataStructures/CustomDataStructures/LinkedList/SinglyLinkedListImplementation/Node.java
@@ -1,0 +1,11 @@
+package DataStructures.CustomDataStructures.LinkedList.SinglyLinkedListImplementation;
+
+public class Node<T> {
+    public T value;
+    public Node<T> next;
+
+    public Node(T newValue) {
+        this.value = newValue;
+        this.next = null;
+    }
+}
