--- src/DataStructures/CustomDataStructures/LinkedList/SinglyLinkedListWithHeadPointer/Node.java
@@ -1,14 +1,0 @@
-package DataStructures.CustomDataStructures.LinkedList.SinglyLinkedListWithHeadPointer;
-
-public class Node<T> {
-
-    public T value;
-    public Node<T> next;
-    public Node<T> head;
-
-    public Node(T value) {
-        this.value = value;
-        this.next = null;
-        this.head = null;
-    }
-}
