+++ test/InterviewProblems/LinkedListProblems/HasCycle/HasCycleUsingHashSetTest.java
@@ -1,7 +1,7 @@
 package InterviewProblems.LinkedListProblems.HasCycle;
 
-import DataStructures.CustomDataStructures.LinkedList.SinglyLinkedList.Node;
-import DataStructures.CustomDataStructures.LinkedList.SinglyLinkedList.SinglyLinkedList;
+import DataStructures.CustomDataStructures.LinkedList.SinglyLinkedListImplementation.Node;
+import DataStructures.CustomDataStructures.LinkedList.SinglyLinkedListImplementation.SinglyLinkedList;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
